back/ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# 기존 분석 시스템 의존성
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.3.7
pinecone-client==2.2.4
sentence-transformers==2.2.2
pymupdf==1.23.8
reportlab==4.0.7
markdown2==2.4.10
pdfkit==1.0.0
requests==2.31.0 